"Working" servo. Lots of stutter.
It seems I cannot get a clean pulse, broadly speaking duty cycles
of divisors of 3 works well, i.e. 33333, 20000, 23333 etc. Others
cause the servo to jitter back and forth.

The Arduino reports nice pulse values for the working dutys, but
values between 2 and 9 for the "bad" dutys.

Next step is writing my own servo control. The plan is to let the
timer trigger on the correct pulse length and keep the rest of the
period low. I may need a "next" buffer to keep the incoming change
and insert it in the following period instead of inserting it into
the current.